+++ dev/loglevel.js
@@ -18,14 +18,14 @@ const broker = new ServiceBroker({
 				//autoPadding: true
 			}
 		},
-		{
+		/*{
 			type: "File",
 			options: {
 				folder: "d:/logs",
 				filename: "moleculer-{date}.log",
 				formatter: "full"
 			}
-		}
+		}*/
 		/*{
 			type: "File",
 			options: {
@@ -84,11 +84,10 @@ const broker = new ServiceBroker({
 				}
 			}
 		},*/
-		/*{
+		{
 			type: "Datadog",
-			options: {
-			}
-		}*/
+			options: {}
+		}
 	],
 	logLevel: {
 		"MY.**": "trace",
@@ -120,11 +119,11 @@ const schema = {
 	}
 };
 
-broker.createService({ ...schema, name: 'greeter', version: 2 });
+broker.createService({ ...schema, name: "greeter", version: 2 });
 
-broker.createService({ ...schema, name: 'test' });
+broker.createService({ ...schema, name: "test" });
 
-broker.createService({ ...schema, name: 'hello' });
+broker.createService({ ...schema, name: "hello" });
 
 const myLogger = broker.getLogger("my.custom.module");