+++ test/unit/metrics/reporters/prometheus.spec.js
@@ -24,6 +24,7 @@ describe("Test Prometheus Reporter class", () => {
 				metricNameFormatter: null,
 				labelNameFormatter: null,
 
+				host: "0.0.0.0",
 				port: 3030,
 				path: "/metrics",
 				defaultLabels: expect.any(Function)
@@ -39,6 +40,7 @@ describe("Test Prometheus Reporter class", () => {
 				metricNameFormatter: () => {},
 				labelNameFormatter: () => {},
 
+				host: "0.0.0.0",
 				port: 12345,
 				path: "/meter"
 			});
@@ -51,6 +53,7 @@ describe("Test Prometheus Reporter class", () => {
 				metricNameFormatter: expect.any(Function),
 				labelNameFormatter: expect.any(Function),
 
+				host: "0.0.0.0",
 				port: 12345,
 				path: "/meter",
 				defaultLabels: expect.any(Function)